AAD token use promise instead of sendSync
