+++ src/client/core/aad/adal/aad.service.ts
@@ -1,6 +1,5 @@
 
-import { BatchLabsApplication } from "client/core";
-import * as ipcMain from "client/core//ipc-promise-main";
+import { BatchLabsApplication, BlIpcMain } from "client/core";
 import { localStorage } from "client/core/local-storage";
 import { logger } from "client/logger";
 import { Constants } from "common";
@@ -15,10 +14,6 @@ import {
 import { AuthenticationService, AuthorizeResult } from "../authentication";
 import { AADUser } from "./aad-user";
 import { UserDecoder } from "./user-decoder";
-
-ipcMain.on("test-async", (data) => {
-    return new Promise((resolve) => setTimeout(() => resolve("Banana"), 1000));
-});
 
 const resources = [
     Constants.ResourceUrl.arm,
