+++ src/client/core/index.ts
@@ -1,4 +1,5 @@
 export * from "./unique-window";
+export * from "./bl-ipc-main";
 export * from "./subprocess";
 export * from "./batchlabs-application";
 export * from "./smart-card-certificate";
